+++ packages/ragbits-document-search/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Update audit imports (#427)
+
 ## 0.17.1 (2025-05-09)
 
 ### Changed